handle local requests for snapshots
