+++ eventkit_cloud/utils/mapproxy.py
@@ -5,13 +5,16 @@ import logging
 import yaml
 from django.conf import settings
 from django.db import connections
+from django.core.cache import cache
 import mapproxy
-from mapproxy.cache.geopackage import GeopackageCache
-from mapproxy.config.config import load_config, load_default_config
-from mapproxy.config.loader import ProxyConfiguration, ConfigurationError, validate_references
 from mapproxy.seed import seeder
 from mapproxy.seed.config import SeedingConfiguration
 from mapproxy.seed.util import ProgressLog, exp_backoff, timestamp, ProgressStore
+from mapproxy.config.config import load_config, load_default_config
+from mapproxy.config.loader import ProxyConfiguration, ConfigurationError, validate_references
+from mapproxy.wsgiapp import MapProxyApp
+
+from webtest import TestApp
 
 import os
 import sqlite3
@@ -311,3 +314,69 @@ def get_concurrency(conf_dict):
     if not concurrency:
         concurrency = getattr(settings, 'MAPPROXY_CONCURRENCY', 1)
     return int(concurrency)
+
+
+def create_mapproxy_app(slug: str):
+    conf_dict = cache.get_or_set(F"base-config-{slug}", lambda: get_conf_dict(slug), 360)
+
+    # TODO: place this somewhere else consolidate settings.
+    base_config = {"services": {"demo": None,
+                                "tms": None,
+                                "wmts": None,
+                                },
+                   "caches": {slug: {"default": {"type": "file"},
+                                     "sources": ["default"],
+                                     "grids": ["default"]}},
+                   "layers": [{"name": slug, "title": slug, "sources": [slug]}],
+                   "globals": {"cache": {"base_dir": getattr(settings, "TILE_CACHE_DIR")}}
+                   }
+    try:
+        mapproxy_config = load_default_config()
+        load_config(mapproxy_config, config_dict=conf_dict)
+        load_config(mapproxy_config, config_dict=base_config)
+        mapproxy_configuration = ProxyConfiguration(mapproxy_config)
+    except ConfigurationError as e:
+        logger.error(e)
+        raise
+
+    cert_var = conf_dict.get("cert_var")
+    auth_requests.patch_https(slug=slug, cert_var=cert_var)
+
+    cred_var = conf_dict.get("cred_var")
+    auth_requests.patch_mapproxy_opener_cache(slug=slug, cred_var=cred_var)
+
+    app = MapProxyApp(mapproxy_configuration.configured_services(), mapproxy_config)
+    return TestApp(app)
+
+
+def get_conf_dict(slug: str) -> dict:
+    """
+    Takes a slug value for a DataProvider and returns a mapproxy configuration as a dict.
+    :param slug: A string matching the slug of a DataProvider
+    :return: a dict.
+    """
+    from eventkit_cloud.jobs.models import DataProvider  # Circular reference
+
+    try:
+        provider = cache.get_or_set(F"DataProvider-{slug}", lambda: DataProvider.objects.get(slug=slug), 360)
+    except Exception:
+        raise Exception(F"Unable to find provider for slug {slug}")
+
+        # Load and "clean" mapproxy config for displaying a map.
+    try:
+        conf_dict = yaml.load(provider.config)
+        conf_dict.pop("caches", "")
+        conf_dict.pop("layers", "")
+        ssl_verify = getattr(settings, "SSL_VERIFICATION", True)
+        if isinstance(ssl_verify, bool):
+            if not ssl_verify:
+                conf_dict['globals'] = {'http': {'ssl_no_cert_checks': ssl_verify}}
+        else:
+            conf_dict['globals'] = {'http': {'ssl_ca_certs': ssl_verify}}
+        conf_dict.update({'globals': {'cache': {'lock_dir': "./locks",
+                                                'tile_lock_dir': "./locks"}}})
+    except Exception as e:
+        logger.error(e)
+        raise Exception(F"Unable to load a mapproxy configuration for slug {slug}")
+
+    return conf_dict
